+++ java/daikon/inv/twoSequence/SuperSequence.java
@@ -42,7 +42,7 @@ class SuperSequence extends TwoSequence {
   }
 
 
-  public void add_modified(int[] a1, int[] a2, int count) {
+  public void add_modified(long[] a1, long[] a2, int count) {
     if (ArraysMDE.indexOf(a1, a2) == -1) {
       destroy();
       return;
